{{cookiecutter.repo_name}}/requirements-dev.txt
autoflake
black
bump2version
coverage
flake8
isort
mypy
pyspark-stubs
s3cmd
tox
tox-gh-actions
